feat: JSON 디코더 확장
